+++ shared/src/index.ts
@@ -10,3 +10,4 @@ export * from './practice-exam';
 export * from './donation-bank';
 export * from './donation-methods';
 export * from './donation-amounts';
+export * from './donation-payment-modes';